headers and regular text inspections

+++ src/main/java/com/prmncr/normativecontrol/dtos/Error.java
@@ -1,4 +1,4 @@
 package com.prmncr.normativecontrol.dtos;
 
-public record Error(long paragraph, long run, ErrorType errorType) {
+public record Error(int paragraph, int run, ErrorType errorType) {
 }
